Archive Critical Review of Gabriel García Márquez archives
I think that I'm doing something wrong with the Markdown format. It seems like it doesn't work.  I follow the tips, but I'm not sure. I have a problem to find the question mark for the beginning of the question. In Spanish we need it. Anyway. This is for far and sure my first draft. Thank you in advance.